app/src/test/java/com/example/theweatherapp/WeatherViewModelTest.kt
package com.example.theweatherapp

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.lifecycle.Observer
import com.example.theweatherapp.model.GeoResponse
import com.example.theweatherapp.model.Main
import com.example.theweatherapp.model.Weather
import com.example.theweatherapp.model.WeatherResponse
import com.example.theweatherapp.repository.WeatherRepo
import com.example.theweatherapp.utils.WeatherContants
import com.example.theweatherapp.viewmodel.WeatherViewModel
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.runTest
import org.junit.Test

import org.junit.Assert.*
import org.junit.Before
import org.junit.Rule
import org.mockito.Mock
import org.mockito.Mockito.verify
import org.mockito.Mockito.`when`
import org.mockito.MockitoAnnotations

class WeatherViewModelTest {

    @Mock
    lateinit var repository: WeatherRepo

    private lateinit var viewModel: WeatherViewModel

    @Mock
    lateinit var weatherObserver: Observer<WeatherResponse>

    @Mock
    lateinit var loadingObserver: Observer<Boolean>

    @Before
    fun setUp() {
        // Initialize Mockito
        MockitoAnnotations.openMocks(this)

        // Initialize the ViewModel with the mocked repository
        viewModel = WeatherViewModel(repository)
    }

    @Test
    fun getWeather_success()  = runTest  {

        assertNotNull(viewModel)

        val fakeMain = Main(25.0)
        val fakeWeather = Weather("Sunny", "")

        val fakeWeatherResponse = WeatherResponse("New York", fakeMain, listOf(fakeWeather))

        // Mock repository response
        `when`(repository.getWeather("New York", WeatherContants.API_KEY, WeatherContants.METRIC))
            .thenReturn(fakeWeatherResponse)

        // Set observers to observe LiveData
        viewModel.weatherData.observeForever(weatherObserver)
        viewModel.loading.observeForever(loadingObserver)

        // Call the function to test
        //viewModel.getWeather("New York")

        // Verify that loading is first set to true and then to false
        //verify(loadingObserver).onChanged(true)
        //verify(loadingObserver).onChanged(false)

        //verify(weatherObserver).onChanged(fakeWeatherResponse)

        //Need to look more into how to mock request written in coroutine.
        //Given more time will do it...
        //val mockResponse = WeatherResponse(name = "London", main = null, weather = null)
    }
}